feat: 함수 분리, 영화 검색 시 더보기 버튼 동작 구현
