+++ types/movie.d.ts
@@ -1,0 +1,22 @@
+export type MovieResponse = {
+  page: number;
+  total_pages: number;
+  results: Movie[];
+};
+
+export type Movie = {
+  adult: boolean;
+  backdrop_path: string;
+  genre_ids: number[];
+  id: number;
+  original_language: string;
+  original_title: string;
+  overview: string;
+  popularity: number;
+  poster_path: string;
+  release_date: string;
+  title: string;
+  video: boolean;
+  vote_average: number;
+  vote_count: number;
+};
